% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_games.R
\name{cfl_game_info}
\alias{cfl_game_info}
\title{\strong{Get results information from games.}}
\source{
\url{http://api.cfl.ca/v1/games/}
}
\usage{
cfl_game_info(season = NA, game_id = NA)
}
\arguments{
\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game - also requires a season}

\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}
}
\value{
\code{\link[=cfl_game_info]{cfl_game_info()}} - A data frame with 14 variables:
\describe{
\item{\code{game_id}: integer.}{Referencing game id.}
\item{\code{date_start}: character.}{Timestamp of kickoff.}
\item{\code{game_number}: integer.}{The sequential league wide game number within the season.}
\item{\code{week}: integer.}{The week of the season.}
\item{\code{season}: integer.}{Season/Year.}
\item{\code{attendance}: integer.}{Recorded number of fans at game.}
\item{\code{event_type}: character.}{The season type: Pre-season, regular season, post-season.}
\item{\code{venue}: character.}{Game location.}
\item{\code{coin_toss}: character.}{Winner of the opening coin toss.}
\item{\code{winner}: character.}{Winning team in the game.}
\item{\code{home_team}: character.}{Home team.}
\item{\code{away_team}: character.}{Away team.}
\item{\code{home_score}: integer.}{Home points.}
\item{\code{away_score}: integer.}{Away points.}
}
}
\description{
\strong{Get results information from games.}
}
\examples{
\donttest{
  cfl_game_info(2019, 2552)
}
}
\keyword{Game}
\keyword{Info}
