% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_games.R
\name{cfl_game_player_stats}
\alias{cfl_game_player_stats}
\title{\strong{Get player statistics by game}}
\source{
\url{http://api.cfl.ca/v1/games/?include=playbyplay}
}
\usage{
cfl_game_player_stats(
  season = NA,
  game_id = NA,
  week = NA,
  team = NA,
  event_type = NA
)
}
\arguments{
\item{game_id}{(\emph{Integer} optional): Game ID filter for querying a single game
Can be found using the \code{\link[=cfl_game_info]{cfl_game_info()}} function}

\item{week}{(\emph{Integer} optional): Week}

\item{team}{(\emph{String} optional): CFL Team Abbreviation}

\item{event_type}{(\emph{String} default regular): Select Season Type: Preseason, Regular Season, Playoffs, Grey Cup}

\item{year}{(\emph{Integer} required): Year, 4 digit format(\emph{YYYY})}
}
\value{
\code{\link[=cfl_game_player_stats]{cfl_game_player_stats()}} - A data frame with 32 variables:
\describe{
\item{\code{game_id} : integer.}{UUID for game}
\item{\code{season} : integer.}{Season game played}
\item{\code{week} : integer.}{Week of Season}
\item{\code{date_start} : character.}{Start date of kickoff}
\item{\code{team} : character.}{Player team}
\item{\code{cfl_central_id} : integer.}{UUID for player}
\item{\code{first_name} : character.}{Player first name}
\item{\code{middle_name} : character.}{Player middle name}
\item{\code{last_name} : character.}{Player last name}
\item{\code{birth_date} : character.}{Player DOB}
\item{\code{pass_attempts} : integer.}{Pass Attempts}
\item{\code{pass_completions} : integer.}{Pass Completios}
\item{\code{pass_net_yards} : integer.}{Net Passing Yards}
\item{\code{pass_long} : integer.}{Long Pass}
\item{\code{pass_touchdowns} : integer.}{Passing TDs}
\item{\code{pass_completion_percentage} : character.}{Completition Percentage}
\item{\code{pass_efficiency} : character.}{Pass Efficiency Rating}
\item{\code{pass_interceptions} : integer.}{Interceptions Thrown}
\item{\code{pass_fumbles} : integer.}{Fumbles while passing}
\item{\code{rush_attempts} : integer.}{Rush attempts}
\item{\code{rush_net_yards} : integer.}{Net rushing yards}
\item{\code{rush_long} : integer.}{Longest rush}
\item{\code{rush_touchdowns} : integer.}{Rushing touchdowns}
\item{\code{rush_long_touchdowns} : integer.}{Longest rushing TD}
\item{\code{receive_attempts} : integer.}{Targets}
\item{\code{receive_caught} : integer.}{Receptions}
\item{\code{receive_yards} : integer.}{Receiving yards}
\item{\code{receive_long} : integer.}{Longest reception}
\item{\code{receive_touchdowns} : integer.}{Receiving TDs}
\item{\code{receive_long_touchdowns} : integer.}{Long receiving TD}
\item{\code{receive_yards_after_catch} : integer.}{Yards After Catch}
\item{\code{receive_fumbles} : integer.}{Fumbles after reception}
\item{\code{punts} : integer.}{Punts}
\item{\code{punt_yards} : integer.}{Punt Yards}
\item{\code{punt_net_yards} : integer.}{Net Punt Yards}
\item{\code{punt_long} : integer.}{Longest Punt}
\item{\code{punt_singles} : integer.}{Punt Singles}
\item{\code{punts_blocked} : integer.}{Punts Blocked}
\item{\code{punts_in_10} : integer.}{Punts downed inside 10}
\item{\code{punts_in_20} : integer.}{Punts downed inside 20}
\item{\code{punts_returned} : integer.}{Punts Returns Defended}
\item{\code{punt_returns} : integer.}{Received Punts Returns}
\item{\code{punt_returns_yards} : integer.}{Punt Return Yards}
\item{\code{punt_returns_touchdowns} : integer.}{Punt Return TDs}
\item{\code{punt_returns_long} : integer.}{Longest Punt Return}
\item{\code{punt_returns_touchdowns_long} : integer.}{Longest Punt Return TD}
\item{\code{kick_returns} : integer.}{Kick Returns}
\item{\code{kick_returns_yards} : integer.}{Kick returns yards}
\item{\code{kick_returns_touchdowns} : integer.}{KR Touchdowns}
\item{\code{kick_returns_long} : integer.}{Longest kick return}
\item{\code{kick_returns_touchdowns_long} : integer.}{Longest kick return TD}
\item{\code{field_goal_attempts} : integer.}{Field goal attempts}
\item{\code{field_goal_made} : integer.}{Field goals made}
\item{\code{field_goal_yards} : integer.}{Attempt distances}
\item{\code{field_goal_singles} : integer.}{Field goal singles}
\item{\code{field_goal_long} : integer.}{Longest field goals}
\item{\code{field_goal_missed_list} : character.}{List of missing distances}
\item{\code{field_goal_points} : integer.}{Points from field goal attempts}
\item{\code{kicks} : integer.}{Kickoffs}
\item{\code{kick_yards} : integer.}{Kickoff yards}
\item{\code{kicks_net_yards} : integer.}{Net kickoff yards}
\item{\code{kicks_long} : integer.}{Longest kickoff}
\item{\code{kicks_singles} : integer.}{Kickoff singles}
\item{\code{kicks_out_of_end_zone} : integer.}{Kicks out of end zone}
\item{\code{kicks_onside} : integer.}{Onside kicks}
\item{\code{one_point_converts_attempts} : integer.}{XP attempts}
\item{\code{one_point_converts_made} : integer.}{XP conversions}
\item{\code{tackles_total} : integer.}{2 pt attempts}
\item{\code{tackles_defensive} : integer.}{2 pt conversions}
\item{\code{tackles_special_teams} : integer.}{Total tackles}
\item{\code{sacks_qb_made} : integer.}{Defensive tackles}
\item{\code{interceptions} : integer.}{Special teams tackles}
\item{\code{fumbles_forced} : integer.}{Sacks made}
\item{\code{fumbles_recovered} : integer.}{Interceptions forced}
\item{\code{passes_knocked_down} : integer.}{Fumbles forced}
\item{\code{fumbles_recovered} : integer.}{Fumble recoveries}
\item{\code{passes_knocked_down} : integer.}{Batted passes}
}
}
\description{
\strong{Get player statistics by game}
}
\examples{
\donttest{
  cfl_game_player_stats(year = 2020, week = 15, team = "BC")

}
}
\keyword{Game}
\keyword{Info}
