% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_games.R
\name{cfl_game_records}
\alias{cfl_game_records}
\title{\strong{Get team records by year}}
\source{
\url{https://api.cfl.ca/v1/standings/}
}
\usage{
cfl_game_records(season = NA, team = NA)
}
\arguments{
\item{team}{(\emph{String} optional): CFL Team Abbreviation}

\item{year}{(\emph{Integer} optional): Year, 4 digit format (\emph{YYYY})}
}
\value{
\code{\link[=cfl_game_records]{cfl_game_records()}} - A data frame with 20 variables:
\describe{
\item{\code{division_id} : integer.}{Division ID}
\item{\code{division_name} : character.}{Division Name}
\item{\code{place} : integer.}{Place in Division}
\item{\code{flags} : character.}{Flags}
\item{\code{team_id} : integer.}{Team Id}
\item{\code{letter} : character.}{First letter of team abbreviation}
\item{\code{abbreviation} : character.}{Team abbreviation}
\item{\code{location} : character.}{Team location}
\item{\code{nickname} : character.}{Team Mascot}
\item{\code{full_name} : character.}{Full name (eg: BC Lions)}
\item{\code{games_played} : integer.}{Games Played}
\item{\code{wins} : integer.}{Wins}
\item{\code{losses} : integer.}{Losses}
\item{\code{ties} : integer.}{Ties}
\item{\code{points} : integer.}{Points in Standings (2pts for W, 1pt for T)}
\item{\code{winning_percentage} : integer.}{Win \%: Pts / (Games Played * 2)}
\item{\code{points_for} : integer.}{Points Scored}
\item{\code{points_against} : integer.}{Points Scored Against}
\item{\code{home_wins} : integer.}{Home Wins}
\item{\code{home_losses} : integer.}{Home Losses}
\item{\code{home_ties} : integer.}{Home Ties}
\item{\code{away_wins} : integer.}{Road wins}
\item{\code{away_losses} : integer.}{Road losses}
\item{\code{away_ties} : integer.}{Road ties}
\item{\code{division_wins} : integer.}{Division Wins}
\item{\code{division_losses} : integer.}{Division Losses}
\item{\code{division_ties} : integer.}{Division Ties}
\item{\code{season} : integer.}{Season}
}
}
\description{
\strong{Get team records by year}
}
\examples{
\donttest{
  cfl_game_records(2018, team = "WPG")

}
}
\keyword{Info}
\keyword{Team}
