% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_games.R
\name{cfl_games}
\alias{cfl_games}
\title{\strong{CFL Games Endpoint Overview}}
\description{
Get results, statistics and information for games\cr
\describe{
\item{\code{cfl_game_player_stats()}:}{ Get results information from games.}
\item{\code{cfl_game_team_stats()}:}{ Get team statistics by game.}
\item{\code{cfl_game_info()}:}{ Get results information from games.}
\item{\code{cfl_game_records()}:}{ Get team records by year.}
\item{\code{cfl_calendar()}:}{ Get calendar of weeks by season.}
}
}
\details{
\preformatted{### **Get player statistics by game**
```r

cfl_game_player_stats(2013, week = 1, team = "WPG")
}
\subsection{\strong{Get team records by year}}{\if{html}{\out{<div class="r">}}\preformatted{cfl_game_records(2018, team = "BC)

}\if{html}{\out{</div>}}
}

\subsection{\strong{Get team statistics by game}}{\if{html}{\out{<div class="r">}}\preformatted{cfl_game_team_stats(2019, team = "CGY")

}\if{html}{\out{</div>}}
}

\subsection{\strong{Get results information from games.}}{\if{html}{\out{<div class="r">}}\preformatted{cfl_game_info(2018, week = 1)

}\if{html}{\out{</div>}}
}

\subsection{\strong{Get calendar of weeks by season.}}{\if{html}{\out{<div class="r">}}\preformatted{cfl_calendar(2019)
}\if{html}{\out{</div>}}
}
}
