% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_pbp_data.R
\name{cfl_pbp_data}
\alias{cfl_pbp_data}
\title{Query for CFL Play-by_Play information}
\usage{
cfl_pbp_data(season = NA, game_id = NA, week = NA, team = NA, event_type = NA)
}
\arguments{
\item{season}{Required parameter for the football season.}

\item{game_id}{Specific Game Id}

\item{week}{Week of the season}

\item{team}{Request for specific team}

\item{event_type}{Request specific event type (Preseason, Regular, Grey Cup, etc)}
}
\value{
\code{\link[=cfl_pbp_data]{cfl_pbp_data()}} - A data frame with 35 variables:
\describe{
\item{\code{game_id} : integer.}{UUID of game}
\item{\code{home_team} : character.}{Home team abbreviation}
\item{\code{away_team} : character.}{Away abbreviation}
\item{\code{season} : integer.}{Season}
\item{\code{week} : integer.}{Week of season}
\item{\code{play_id} : integer.}{UUID of play}
\item{\code{play_sequence} : integer.}{Play number of game}
\item{\code{quarter} : integer.}{Quarter}
\item{\code{play_clock_start} : character.}{Play clock at snap}
\item{\code{play_clock_start_in_secs} : integer.}{Seconds rem in quarter at snap}
\item{\code{field_position_start} : character.}{Field position start}
\item{\code{field_position_end} : character.}{Field position end}
\item{\code{down} : integer.}{Down}
\item{\code{yards_to_go} : integer.}{Yards for first down}
\item{\code{is_in_red_zone} : logical.}{Is inside 20}
\item{\code{team_home_score} : integer.}{Home score}
\item{\code{team_visitor_score} : integer.}{Away score}
\item{\code{play_type_id} : integer.}{Play Type Id}
\item{\code{play_result_type_id} : integer.}{Result Type Id}
\item{\code{play_result_yards} : integer.}{Result Yards}
\item{\code{play_result_points} : integer.}{Result Points}
\item{\code{play_success_id} : integer.}{Success Id}
\item{\code{play_change_of_possession_occurred} : logical.}{Last play of posession}
\item{\code{team_abbreviation} : character.}{Offense Abbreviation}
\item{\code{team_id} : integer.}{Offensive Team Id}
\item{\code{play_summary} : character.}{Play Summary}
\item{\code{distance_to_goal} : numeric.}{Yards to Goal}
\item{\code{seconds_remain_half} : numeric.}{Seconds rem in  half at snap}
\item{\code{seconds_remain_game} : numeric.}{Seconds rem in  game at snap}
\item{\code{quarterback_id} : integer.}{Quarterback CFL Player Id}
\item{\code{ball_carrier_id} : integer.}{Ball carrier CFL player Id}
\item{\code{primary_defender_id} : integer.}{Primary defender CFL Player Id}
\item{\code{play_result_type_desc} : character.}{Play Result Type}
\item{\code{play_type_desc} : character.}{Play Type}
\item{\code{play_success_desc} : character.}{Success Description}
}
}
\description{
Query for CFL Play-by_Play information
}
\examples{
\dontrun{
cfl_pbp_data(season = 2016, game_id = 2280)
}
}
