% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfl_pbp_data.R
\name{cfl_plays}
\alias{cfl_plays}
\alias{cfl_pbp_fetch}
\title{\strong{CFL Games Endpoint Overview}}
\usage{
cfl_pbp_fetch(season = NA, game_id = NA)
}
\description{
Get results, statistics and information for games\cr
\describe{
\item{\code{cfl_pbp_data()}:}{ Get play by play data from games.}
}
}
\details{
###** Get PxP Data by year **\if{html}{\out{<div class="r">}}\preformatted{bc_pbp <- cfl_pbp_data(season = 2018, team = 'BC')


}\if{html}{\out{</div>}}
}
