% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfbfastR-package.R
\docType{package}
\name{cflfastR-package}
\alias{cflfastR}
\alias{cflfastR-package}
\title{cflfastR: Wrapping the official CFL Data API to provide clean-ish play by play data.}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

cflfastR is a wrapper for the CFL API to provide play by play data.
}
\author{
\strong{Maintainer}: Garrett McClintock \email{garrettjmcclintock@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
